refactor(PlayerMoney): 플레이어 베팅금액 관리하는 일급컬렉션 추가

+++ src/main/java/blackjack/domain/game/Money.java
@@ -6,7 +6,7 @@ public class Money {
 
     private final double value;
 
-    public Money(int value) {
+    public Money(double value) {
         this.value = value;
     }
